dataproc/requirements.txt
flask
requests
numpy
pandas
cassandra-driver
python-dotenv
astrapy
pyspark
google-cloud-storage
google-cloud-dataproc
